+++ AssetsFlow.Server/AssetsFlow.Web.API/Configurations/CacheConfig.cs
@@ -1,28 +1,52 @@
-using HsR.Web.API.Settings;
+using AssetsFlowWeb.Services.Models;
+using AssetsFlowWeb.Services.Models.Journal;
+using HsR.Common.Services.Caching;
+using HsR.Journal.Entities;
+using HsR.UserService.Protos;
 using HsR.Web.API.Services;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+using HsR.Web.API.Settings;
+using Microsoft.Extensions.Caching.Memory;
 
-namespace AssetsFlowWeb.API.Configurations
+public static class ServiceCollectionExtensions
 {
-    public static class CacheConfig
+    public static IServiceCollection AddCacheServices(this IServiceCollection services, IConfiguration configuration)
     {
-        public static IServiceCollection AddCacheServices(this IServiceCollection services, IConfiguration configuration)
+        var cacheSettings = configuration.GetSection("CacheSettings").Get<CacheSettings>() ?? new CacheSettings();
+
+        services.AddMemoryCache(options =>
         {
-            var cacheSettings = configuration.GetSection("CacheSettings").Get<CacheSettings>() ?? new CacheSettings();
-            services.AddMemoryCache(options =>
-            {
-                options.SizeLimit = cacheSettings.SizeLimit;
-            });
-            services.AddSingleton<ITradesCacheService, TradesCacheService>();
-            services.AddHostedService<CacheCleanupService>(provider =>
-                new CacheCleanupService(
-                    provider.GetRequiredService<ITradesCacheService>(),
-                    provider.GetRequiredService<Microsoft.Extensions.Logging.ILogger<CacheCleanupService>>(),
-                    provider.GetRequiredService<IConfigurationService>()
-                )
+            options.SizeLimit = cacheSettings.SizeLimit;
+        });
+
+        // Register cache services
+        services.AddSingleton<ICacheService<Guid, IEnumerable<TradeCompositeModel>>, TradesCacheService>();
+        services.AddSingleton<ICacheService<Guid, UserDataDTO>, UserDataCacheService>();
+
+        // Register cleanup background services
+        services.AddHostedService<CacheCleanupService<Guid, List<TradeCompositeModel>>>(provider =>
+        {
+            var cache = provider.GetRequiredService<ICacheService<Guid, List<TradeCompositeModel>>>();
+            var logger = provider.GetRequiredService<ILogger<CacheCleanupService<Guid, List<TradeCompositeModel>>>>();
+            return new CacheCleanupService<Guid, List<TradeCompositeModel>>(
+                cache,
+                logger,
+                TimeSpan.FromMinutes(cacheSettings.CleanupIntervalMinutes),
+                TimeSpan.FromHours(cacheSettings.CleanupInactiveUsersThresholdHours)
             );
-            return services;
-        }
+        });
+
+        services.AddHostedService<CacheCleanupService<Guid, UserDto>>(provider =>
+        {
+            var cache = provider.GetRequiredService<ICacheService<Guid, UserDto>>();
+            var logger = provider.GetRequiredService<ILogger<CacheCleanupService<Guid, UserDto>>>();
+            return new CacheCleanupService<Guid, UserDto>(
+                cache,
+                logger,
+                TimeSpan.FromMinutes(cacheSettings.CleanupIntervalMinutes),
+                TimeSpan.FromHours(cacheSettings.CleanupInactiveUsersThresholdHours)
+            );
+        });
+
+        return services;
     }
-} 
+}