+++ AssetsFlow.Server/CHANGELOG.md
@@ -4,6 +4,8 @@
 - filtering: added to userdata
 - userdata: added repo wrapper for get symbols, get filters
 - closure: change lessons to : mistakes, strengths, takeaways
+- morphed cache service into generic
+- added userdata to cache
 
 ## [0.9.225-b] - 2025-08-18
 - validate net amount on close trade, return reduce element on close